计算机视觉导论/super-resolution/SISR/requirements.txt
torch
torchvision
pillow
numpy
tqdm
pyyaml
scikit-image
thop
pandas
matplotlib